Added Permaculture Business Ecosystem Visualization
This Document visualizes the permacutlutre business ecosystem I have in mind. 
It demonstrates the potential lying inside the simplification of  the permaculture design process through free and open source, smart and ai supported planning software.
I also tried to visualize all the market potential that is created through the creation of this ecosystem, demonstrating that this concept can carry it's own wheight easily and that it is actually a way of creating potential markets and high value jobs where there are none at the moment.